Update patch set 1
Patch Set 1:

Sorry, this is a rather huge review. This is because i had to implement a lot of features first, to test if implementing an own data-grid is possible.
I tried to remove some of them to reduce the ticket size, but splitting it all down into small subtasks would have been to time consuming.
If you would like, we can do the review together. Or i can give you a short introduction to the code.

Patch-set: 1
Reviewer: Gerrit User 1013694 <1013694@d5d70762-12d0-45a1-890d-524b12d3f735>