Update patch set 1
Patch Set 1:

In this commit i moved the samples grid ng-template definitions to the cell components. So the components itself are not in the dom anymore. This reduces the dom-element count for each cell by one.

Additionally the template inputs for the data-grid are not defined for each cell anymore. They are now selected by an id in the cell view model.

Patch-set: 1
Reviewer: Gerrit User 1013694 <1013694@d5d70762-12d0-45a1-890d-524b12d3f735>